split tests into a few different files (arbitrarily)
